requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click-plugins==1.1.1
click==8.1.6
cligj==0.7.2
colorama==0.4.6
customtkinter==5.2.0
darkdetect==0.8.0
Fiona==1.9.4.post1
geopandas==0.13.2
idna==3.4
networkx==3.1
numpy==1.25.2
osmnx==1.6.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pip==23.2.1
pyproj==3.6.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
setuptools==68.0.0
shapely==2.0.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
wheel==0.41.0
